working article list and functioning editor
